rename module to load-balancer-operator
