+++ .golangci.yml
@@ -1,6 +1,6 @@
 linters-settings:
   goimports:
-    local-prefixes: go.infratographer.com/loadbalanceroperator
+    local-prefixes: go.infratographer.com/load-balancer-operator
 
 linters:
   enable:
